# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  acceso.AccesoRequest:
    properties:
      nombre:
        type: string
      ubicacion:
        type: string
    required:
    - nombre
    type: object
  acceso.AccesoResponse:
    properties:
      clave_kiosko:
        type: string
      id:
        type: integer
      nombre:
        type: string
      ubicacion:
        type: string
      usuario_id:
        type: integer
    type: object
  admin.AdminRequest:
    properties:
      apellido_materno:
        type: string
      apellido_paterno:
        type: string
      correo:
        type: string
      nombre:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - apellido_materno
    - apellido_paterno
    - correo
    - nombre
    - password
    type: object
  admin.AdminResponse:
    properties:
      apellido_materno:
        type: string
      apellido_paterno:
        type: string
      correo:
        type: string
      id:
        type: integer
      nombre:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend del proyecto AUTOnomIA para la FEPRO 2026 de la BUAP. Gestiona
    el registro y control de visitantes/accesos en condominios.
  title: AUTOnomIA API
  version: "1.0"
paths:
  /accesos:
    get:
      description: Devuelve todos los accesos del admin autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/acceso.AccesoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar accesos
      tags:
      - accesos
    post:
      consumes:
      - application/json
      description: Registra un nuevo acceso para el admin y genera su clave de kiosko
        (se devuelve en texto plano una sola vez, aqui)
      parameters:
      - description: Datos del acceso
        in: body
        name: acceso
        required: true
        schema:
          $ref: '#/definitions/acceso.AccesoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/acceso.AccesoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear acceso
      tags:
      - accesos
  /accesos/{id}:
    delete:
      description: Elimina un acceso del admin autenticado
      parameters:
      - description: ID del acceso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar acceso
      tags:
      - accesos
    get:
      description: Busca un acceso del admin autenticado por su ID
      parameters:
      - description: ID del acceso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/acceso.AccesoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener acceso por ID
      tags:
      - accesos
    patch:
      consumes:
      - application/json
      description: Modifica el nombre o ubicacion de un acceso del admin autenticado
      parameters:
      - description: ID del acceso
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a modificar
        in: body
        name: acceso
        required: true
        schema:
          $ref: '#/definitions/acceso.AccesoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/acceso.AccesoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Modificar acceso
      tags:
      - accesos
...
  /admins/{id}:
    delete:
      description: Elimina un administrador
      parameters:
      - description: ID del admin
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar admin
      tags:
      - admins
    get:
      description: Busca un administrador por su ID
      parameters:
      - description: ID del admin
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener admin por ID
      tags:
      - admins
    patch:
      consumes:
      - application/json
      description: Modifica los datos de un administrador, re-hasheando su password
      parameters:
      - description: ID del admin
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a modificar
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.AdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Modificar admin
      tags:
      - admins
...
swagger: "2.0"